Move proposals into separate app
