+++ presenters/admin.py
@@ -1,6 +1,5 @@
 from django.contrib import admin
 from presenters.models.presenter import Presenter
-from presenters.models.proposal import Proposal, Attachment
 
 
 class PresenterAdmin(admin.ModelAdmin):
@@ -19,5 +18,3 @@ class PresenterAdmin(admin.ModelAdmin):
 
 
 admin.site.register(Presenter, PresenterAdmin)
-admin.site.register(Proposal)
-admin.site.register(Attachment)
