+++ presenters/models/__init__.py
@@ -1,2 +1,1 @@
 from .presenter import Presenter
-from .proposal import Proposal, Attachment
