+++ proposals/views/cfp.py
@@ -1,0 +1,24 @@
+from django.views.generic import View
+from django.shortcuts import render
+from django.contrib.auth.decorators import login_required
+from django.utils.decorators import method_decorator
+
+from ..forms import CFPForm
+
+
+class CFPView(View):
+    template_name = 'cfp.html'
+    form_class = CFPForm
+
+    def get(self, request, *args, **kwargs):
+        form = self.form_class()
+        return render(request, self.template_name, {'form': form})
+
+    @method_decorator(login_required)
+    def post(self, request, *args, **kwargs):
+        form = self.form_class(request.POST)
+
+        if form.is_valid():
+            pass
+
+        return render(request, self.template_name, {'form': form})
