+++ pyconlt/settings/base.py
@@ -48,6 +48,7 @@ INSTALLED_APPS = [
     # project apps
     'info',
     'presenters',
+    'proposals'
 ]
 
 MIDDLEWARE = [
